Refactor Checkout
